+++ src/index.tsx
@@ -11,8 +11,9 @@ import Routes from "@/routes";
 import { GoogleOAuthProvider } from "@react-oauth/google";
 
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { StyledToast } from "./styles/toast.styes";
 import { SpeedInsights } from "@vercel/speed-insights/react";
+import { AppProviders } from "./providers";
+import { StyledToast } from "./styles/toast.styes";
 
 const queryClient = new QueryClient();
 
@@ -24,9 +25,11 @@ ReactDOM.createRoot(document.getElementById("root")!).render(
         <StyledToast />
         <SpeedInsights />
         <GlobalStyle theme={darkTheme} />
-        <AppContainer className="flex_ssr">
-          <Routes />
-        </AppContainer>
+        <AppProviders>
+          <AppContainer className="flex_ssr">
+            <Routes />
+          </AppContainer>
+        </AppProviders>
       </GoogleOAuthProvider>
     </QueryClientProvider>
   </React.StrictMode>
