--- src/interfaces/user.ts
@@ -1,7 +1,0 @@
-import { User } from "@/types";
-
-export interface UserInterface {
-  get(): Promise<void>;
-  post(data: User): Promise<User>;
-  getByEmail(email: string): Promise<User>;
-}
