+++ src/providers/interfaces/auth.ts
